xf_compression: final update to (2020.2) - bugfixes & enhancements

+++ data_compression/L2/tests/src/lz4_stream.hpp
@@ -18,22 +18,23 @@
  * @file xil_lz4.hpp
  * @brief Header for LZ4 host functionality
  *
- * This file is part of Vitis Data Compression Library host code for lz4 compression.
+ * This file is part of Vitis Data Compression Library host code for lz4
+ * compression.
  */
 
 #ifndef _XFCOMPRESSION_LZ4_STREAM_HPP_
 #define _XFCOMPRESSION_LZ4_STREAM_HPP_
 
+#include "xcl2.hpp"
 #include <assert.h>
+#include <fstream>
 #include <iomanip>
 #include <iostream>
-#include <stdint.h>
-#include <vector>
 #include <math.h>
-#include <time.h>
+#include <stdint.h>
 #include <string>
-#include <fstream>
-#include "xcl2.hpp"
+#include <time.h>
+#include <vector>
 // This extension file is required for stream APIs
 #include <CL/cl_ext_xilinx.h>
 
@@ -236,8 +237,8 @@ class xfLz4Streaming {
     std::vector<uint8_t, aligned_allocator<uint8_t> > h_buf_in;
     std::vector<uint8_t, aligned_allocator<uint8_t> > h_buf_out;
     std::vector<uint32_t, aligned_allocator<uint32_t> > h_buf_decompressSize;
-    std::vector<uint32_t, aligned_allocator<uint8_t> > h_blksize;
-    std::vector<uint32_t, aligned_allocator<uint8_t> > h_compressSize;
+    std::vector<uint32_t, aligned_allocator<uint32_t> > h_blksize;
+    std::vector<uint32_t, aligned_allocator<uint32_t> > h_compressSize;
 
     // Device buffers
     cl::Buffer* buffer_input;
